client/requirements_django.txt
Django==5.2.8
psycopg2-binary==2.9.10
python-dotenv==1.2.1
gunicorn==23.0.0
openai==2.8.1
requests==2.32.3
grpcio==1.70.0
protobuf==5.29.3
pydicom>=2.4.4
numpy>=1.26.4
scipy>=1.12.0
scikit-image>=0.25.0
matplotlib>=3.8.0
opencv-python>=4.10.0
Pillow>=11.0.0
pytz==2025.1
tzdata==2025.1
